move tests of application-header component to spec of application-banner component
